requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database
pymongo==4.5.0
python-dotenv==1.0.0

# Computer Vision and AI
opencv-python==4.8.1.78
dlib==19.24.2
numpy==1.24.3
imutils==0.5.4

# Audio Processing
pyaudio==0.2.11

# System and Utilities
winsound; sys.platform == "win32"
datetime

# Additional AI/ML Libraries
scikit-learn==1.3.0
scipy==1.11.1

# Development and Testing
pytest==7.4.2
pytest-cov==4.1.0

# Security
bcrypt==4.0.1
cryptography==41.0.4

# HTTP Requests
requests==2.31.0

# Data Processing
pandas==2.0.3

# Image Processing
Pillow==10.0.1

# YOLO Object Detection Dependencies
torch==2.0.1
torchvision==0.15.2

# Face Recognition
face-recognition==1.3.0

# Additional Utilities
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2 